Meeting with Phil on 03/03/2025.
Checked over viabilityMask - made sure everything is working okay. Changed 1:9 to 1:size to remove the possibility of errors. Future task, if have time, could be to combine the two lines of viabilityMask into one line of code. Removed old MEDEA script lines which were there for helpful comparison.

Sex Modifiers - moved onto next task which is to set a parameter that a particular genotype is a specific sex (something that is not necessary for MEDEA but needed for MEREA due to sex-linked genotypes). My task is to make a named vector where each genotype is 0 or 1 and also define phi (see hashtag notes).

return (list()) - changed MEDEA genotypes to our chosen MEREA wildType and releaseType alleles.

After completing tasks, we are hoping to run a simulation without errors.
An auxiliary run is needed before running a simulation.